Move river client closure to caller
